+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Add `LiabilityShift` class and `liability_shift` to RiskData
 * Add ExchangeRateQuote API
+* Allow vaulting of raw ApplePayCards with billing address via Customer.create/update
 
 ## 4.15.2
 * Add `retried` to `Transaction`