+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 4.16.0
 * Add `LiabilityShift` class and `liability_shift` to RiskData
 * Add ExchangeRateQuote API
 * Add `ach_return_responses_created_at` and `reason_code` fields in TransactionSearch